PIM-10439: move directory
