+++ src/Akeneo/Pim/Enrichment/Bundle/Command/Category/CheckCategoryTrees/CheckCategoryTrees.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\Command\Category;
+namespace Akeneo\Pim\Enrichment\Bundle\Command\Category\CheckCategoryTrees;
 
-use Akeneo\Tool\Bundle\ClassificationBundle\Doctrine\ORM\Repository\CategoryRepository;
 use Doctrine\DBAL\Connection;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
