+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/command.yml
@@ -61,7 +61,7 @@ services:
         tags:
             - { name: console.command }
 
-    Akeneo\Pim\Enrichment\Bundle\Command\Category\CheckCategoryTrees:
+    Akeneo\Pim\Enrichment\Bundle\Command\Category\CheckCategoryTrees\CheckCategoryTrees:
         arguments:
             - '@database_connection'
         tags:
